+++ dot_config/quickshell/end4-pC/modules/ii/dynamicIsland/MusicActivity.qml
@@ -1,4 +1,5 @@
 import QtQuick
+import qs.services
 import QtQuick.Layouts
 import Quickshell.Widgets